# specifications/openapi/omlox-hub.v0.yaml
openapi: 3.1.0
info:
  title: Open Location Hub API
  version: 0.2.0
  description: |
    REST API documentation for Open Location Hub.

    This API provides resource management for zones, trackables, location
    providers, and fences, together with provider ingest endpoints and a
    JSON-RPC control-plane interface.

    Parts of this API follow the omlox v2 hub model and terminology where that
    improves interoperability with omlox-compatible systems.

    This contract is independently authored and is not an official PI/PNO
    publication or certification artifact. `omlox` is a trademark of
    PROFIBUS Nutzerorganisation e.V.
...
servers:
- url: http://localhost:8080
security:
- bearerAuth: []
...
paths:
...
  /v2/providers/{providerId}/trackables:
    parameters:
    - $ref: '#/components/parameters/ProviderId'
    get:
      tags:
      - providers
      operationId: getProviderTrackables
      summary: List associated trackables
      description: Returns trackables associated with the provider.
      responses:
        '200':
          description: Matching resources.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trackable'
        '404':
          $ref: '#/components/responses/NotFound'
  /v2/fences/{fenceId}/trackables:
    parameters:
    - $ref: '#/components/parameters/FenceId'
    get:
      tags:
      - fences
      operationId: getFenceTrackables
      summary: List trackables inside the fence
      description: Returns resources currently inside the fence. By default membership includes active exit
        timers; spatial_query=true uses geometry only.
      responses:
        '200':
          description: Matching resources.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trackable'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - $ref: '#/components/parameters/QuerySpatialQuery'
  /v2/fences/{fenceId}/motions:
    parameters:
    - $ref: '#/components/parameters/FenceId'
    get:
      tags:
      - fences
      operationId: getFenceMotions
      summary: List motions inside the fence
      description: Returns resources currently inside the fence. By default membership includes active exit
        timers; spatial_query=true uses geometry only.
      responses:
        '200':
          description: Matching resources.
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                  - $ref: '#/components/schemas/TrackableMotion'
                  - $ref: '#/components/schemas/GeoJSONFeatureCollection'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - $ref: '#/components/parameters/QueryCrs'
      - $ref: '#/components/parameters/QueryZoneId'
      - $ref: '#/components/parameters/QueryGeojson'
      - $ref: '#/components/parameters/QuerySpatialQuery'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentication uses a JWT bearer token. Authorization is enforced by the
        hub's server-side permission registry instead of OAuth
        scope declarations in this OpenAPI document.
  parameters:
...
    ProviderId:
      in: path
      name: providerId
      required: true
      description: Stable provider identifier chosen by the client.
      schema:
        type: string
    FenceId:
      in: path
      name: fenceId
      required: true
      description: UUID of the fence resource.
      schema:
        type: string
        format: uuid
    QueryCrs:
      name: crs
      in: query
      description: Output coordinate reference system. Defaults to EPSG:4326; local returns original
        coordinates unless zone_id selects another zone.
      schema:
        type: string
    QueryZoneId:
      name: zone_id
      in: query
      description: Target zone for local coordinate projection. Used with crs=local.
      schema:
        type: string
    QueryGeojson:
      name: geojson
      in: query
      description: Return GeoJSON features instead of resource envelopes.
      schema:
        type: boolean
...
    QuerySpatialQuery:
      name: spatial_query
      in: query
      description: Use current geometric containment instead of membership maintained by fence timers.
      schema:
        type: boolean
  responses:
...
    NotFound:
      description: The requested resource identifier does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      description: Standard error envelope returned by the REST API.
      required: [type, code]
      properties:
        type:
          type: string
          description: Stable machine-readable error type.
        code:
          type: integer
          description: HTTP status code associated with the error.
        message:
          type: string
          description: Human-readable error message.
        details:
          type: array
          description: Actionable validation, routing, or authorization details that explain how to correct
            the request.
          items:
            type: string
    ExtensionProperties:
      type: object
      description: Free-form extension object preserved as-is by the hub.
      additionalProperties: true
    StringIdList:
      type: array
      description: List of opaque string identifiers.
      items:
        type: string
    GeoJsonPosition2D:
      type: array
      description: GeoJSON 2D coordinate tuple in `[x, y]` order.
      minItems: 2
      maxItems: 2
      items:
        type: number
        format: double
    GeoJsonPosition3D:
      type: array
      description: GeoJSON 3D coordinate tuple in `[x, y, z]` order.
      minItems: 3
      maxItems: 3
      items:
        type: number
        format: double
    GeoJsonPosition:
      description: GeoJSON coordinate tuple used by `Point`, `LineString`, and `Polygon`.
      oneOf:
      - $ref: '#/components/schemas/GeoJsonPosition2D'
      - $ref: '#/components/schemas/GeoJsonPosition3D'
    PositiveNumber:
      type: number
      description: Finite nonnegative numeric value. Zero is allowed.
      minimum: 0

      format: double
    PositiveOrMinusOne:
      description: Nonnegative numeric value or `-1` for an infinite timeout. Zero means immediate.
      anyOf:
      - type: number
        enum: [-1]
        format: double
      - $ref: '#/components/schemas/PositiveNumber'
    GeoJSONFeatureCollection:
      type: object
      description: OMLOX GeoJSON wrapper, returned when geojson=true.
      required: [type, features]
      properties:
        type:
          type: string
          const: FeatureCollection
        features:
          type: array
          items:
            type: object
            required: [type, geometry, properties]
            properties:
              type:
                type: string
                const: Feature
              geometry:
                oneOf:
                - $ref: '#/components/schemas/Point'
                - $ref: '#/components/schemas/Polygon'
                - $ref: '#/components/schemas/LineString'
              properties:
                type: object
                additionalProperties: true
    Point:
      type: object
      description: GeoJSON point geometry.
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Point
          description: GeoJSON type discriminator.
        coordinates:
          $ref: '#/components/schemas/GeoJsonPosition'
    LineString:
      type: object
      description: GeoJSON line geometry.
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: LineString
        coordinates:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/GeoJsonPosition'
    Polygon:
      type: object
      description: GeoJSON polygon geometry. The first and last point of each ring are expected to be equal.
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Polygon
        coordinates:
          type: array
          minItems: 1
          items:
            type: array
            minItems: 4
            items:
              $ref: '#/components/schemas/GeoJsonPosition'
...
    LocatingRule:
      type: object
      description: Trackable locating rule used to prioritize candidate locations from providers.
      required: [expression, priority]
      properties:
        expression:
          type: string
          description: >
            Boolean expression using comparisons, AND, parentheses, and literals.
            Supported properties are accuracy, name (provider name), provider_id,
            type (provider technology), source, floor, speed, and timestamp_diff
            (generated-time age in milliseconds). Missing comparison operands do
            not match; omitted floor is zero. Nesting is limited to 64 groups.
          maxLength: 4096
        priority:
          type: number
          description: Higher positive values take precedence. Unmatched locations have priority zero; ties
            use the newest timestamp_generated.

          minimum: 0
          format: double
...
    Trackable:
      type: object
      description: Stored trackable resource returned by the REST API.
      required: [id, type]
      properties:
        id:
          type: string
          format: uuid
          description: Hub-generated trackable identifiers are UUIDv7 when the hub creates them.
        type:
          type: string
          enum: [omlox, virtual]
        name:
          type: string
        geometry:
          $ref: '#/components/schemas/Polygon'
        extrusion:
          type: number
          minimum: 0
          description: Height in meters. Zero represents a plane; omitted values use the floor or unbounded
            height.
          format: double
        location_providers:
          $ref: '#/components/schemas/StringIdList'
          description: Preferred providers considered when deriving trackable locations.
        fence_timeout:
          $ref: '#/components/schemas/PositiveOrMinusOne'
        exit_tolerance:
          $ref: '#/components/schemas/PositiveNumber'
        tolerance_timeout:
          $ref: '#/components/schemas/PositiveOrMinusOne'
        exit_delay:
          $ref: '#/components/schemas/PositiveOrMinusOne'
        radius:
          type: number
          description: |
            Circular extent in meters used for fence and collision decisions.
            Defaults to zero. When set, motion geometry approximates this circle.
          minimum: 0
          default: 0
          format: double
        properties:
          $ref: '#/components/schemas/ExtensionProperties'
        locating_rules:
          type: array
          description: Ordered locating rules applied when multiple candidate locations exist.
          items:
            $ref: '#/components/schemas/LocatingRule'
...
    Location:
      type: object
      description: Provider location observation accepted by ingest and reused in derived outputs.
      required: [position, source, provider_type, provider_id]
      properties:
        position:
          $ref: '#/components/schemas/Point'
        source:
          type: string
          description: Source locator or sensor identifier that produced the observation.
        provider_type:
          x-go-type: string
          type: string
          description: Provider technology or implementation type.
          enum:
          - uwb
          - gps
          - wifi
          - rfid
          - ibeacon
          - virtual
          - unknown
        provider_id:
          type: string
          description: Stable provider identifier.
        trackables:
          $ref: '#/components/schemas/StringIdList'
          description: Trackables associated with the observation, when already known.
        timestamp_generated:
          type: string
          format: date-time
          description: Timestamp when the provider generated the observation.
        timestamp_sent:
          type: string
          format: date-time
          description: Timestamp when the provider sent the observation to the hub.
        crs:
          type: string
          description: Valid EPSG identifier or `local`. Omitted values are treated as `local` by the current
            implementation.
        associated:
          type: boolean
          description: Indicates whether the location has already been associated with a trackable.
        accuracy:
          type: number
          description: Estimated positional accuracy in meters.
          format: double
        floor:
          type: number
          description: Optional floor number or level indicator.
          format: double
        true_heading:
          type: number
          description: Heading in degrees relative to true north.
          format: double
        magnetic_heading:
          type: number
          description: Heading in degrees relative to magnetic north.
          format: double
        heading_accuracy:
          type: number
          description: Estimated heading accuracy in degrees.
          format: double
        elevation_ref:
          type: string
          enum: [floor, wgs84]
          description: Reference system for altitude-like values carried in the geometry.
        speed:
          type: number
          description: Estimated speed in meters per second. Derived hub outputs may also populate this from
            normalized track movement.
          format: double
        course:
          type: number
          description: Travel course in degrees. Derived hub outputs may also populate this from normalized
            track movement.
          format: double
        properties:
          $ref: '#/components/schemas/ExtensionProperties'
...
    TrackableMotion:
      type: object
      description: Derived trackable motion payload published by the hub.
      required: [id, location]
      properties:
        id:
          type: string
        name:
          type: string
        extrusion:
          type: number
          format: double
        geometry:
          $ref: '#/components/schemas/Polygon'
        location:
          $ref: '#/components/schemas/Location'
        properties:
          $ref: '#/components/schemas/ExtensionProperties'
...